+++ src/interfaces/resources/event-artist.interface.ts
@@ -1,11 +1,12 @@
 import { ArtistInterface, createArtist } from "./artist.interface";
 
 export interface EventArtistInterface {
-    id?: string;
+    readonly id?: string;
     event_id: string;
     artist: ArtistInterface,
     rank: number;
     set_time: Date;
+    importance: number;
     readonly created_at?:Date;
     readonly updated_at?:Date;
 }
@@ -21,6 +22,7 @@ export const createEventArtist = (base: any = {}): EventArtistInterface => {
 			artist: {},
 			rank: 0,
 			set_time: "",
+			importance: 0,
 		},
 		...base
 	}  as EventArtistInterface;